% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_outliers.R
\name{visualize_outliers}
\alias{visualize_outliers}
\title{A function that plot out the given data with outliers marked.}
\usage{
visualize_outliers(dataframe, columns = NULL, type = "violin")
}
\arguments{
\item{dataframe}{A target dataframe where the function is performed.}

\item{columns}{The target columns where the function needed to be performed. Default is None, the function will check all columns.}

\item{type}{The method of plotting the distribution.}
}
\value{
A ggplot of data distribution.
}
\description{
A function that plot out the given data with outliers marked.
}
\examples{

library(tidyverse)
library(ggplot2)

df <- tibble(cola=c(1:5), colb=c(6:10), colc=c(11:15))
visualize_outliers(df, columns=c("cola", "colb"), type="violin")
}
